DRY (Don't Repeat Yourself) code - refactoring and partials
DRY - Don't Repeat Yourself - code can be accomplished by refactoring or 'extracting away code redundancies'. We accomplished a significant amount of this by using a couple of private methods in our articles controller and using partials in our views. Clean code should always be a priority once features are built and working as we want so that our code is easy to read and easily maintainable. All changes we made to the application during the video can be found here:

https://github.com/udemyrailscourse/alpha-blog-6/commit/b36fdad0dfb852d07c59b0c7b04e810cd13778e0